fix/moved static data for Q&A to the folder to Q&A
